artist_wiki/requirements.txt
django==4.1.4
djangorestframework==3.14.0
django-filter==22.1

PyMySQL==1.0.2